Homeworks 0 Exercise 1 Documents Style Change

+++ Homeworks/0_cpp_warmup/documents/1_BasicDArray/C2Cpp/src/Array5/Array.h
@@ -8,33 +8,33 @@ template <class DataType>
 class DArray
 {
 private:
-	int n; // the size of the array
-	DataType *pData;  // the data of the array
+	int n;			 // the size of the array
+	DataType *pData; // the data of the array
 
 public:
 	DArray();
 	~DArray();
-	int		InitArray( );
-	int		SetArraySize( int size );
-	int		FreeArray( );
-	int		SetValue( int k, DataType value );
-	int		PrintArray(  );
+	int InitArray();
+	int SetArraySize(int size);
+	int FreeArray();
+	int SetValue(int k, DataType value);
+	int PrintArray();
 };
 
-template<class DataType>
+template <class DataType>
 DArray<DataType>::DArray()
 {
-	InitArray( );
+	InitArray();
 }
 
-template<class DataType>
+template <class DataType>
 DArray<DataType>::~DArray()
 {
-	FreeArray( );
+	FreeArray();
 }
 
-template<class DataType>
-int		DArray<DataType>::InitArray( )
+template <class DataType>
+int DArray<DataType>::InitArray()
 {
 	n = 0;
 	pData = NULL;
@@ -42,12 +42,12 @@ int		DArray<DataType>::InitArray( )
 	return 1;
 }
 
-template<class DataType>
-int		DArray<DataType>::SetArraySize(  int size )
+template <class DataType>
+int DArray<DataType>::SetArraySize(int size)
 {
 	n = size;
-	pData = (DataType*)malloc( size * sizeof(DataType) );
-	if( pData == NULL )
+	pData = (DataType *)malloc(size * sizeof(DataType));
+	if (pData == NULL)
 	{
 		printf("no enough memory!\n");
 		return 0;
@@ -56,48 +56,47 @@ int		DArray<DataType>::SetArraySize(  int size )
 	return 1;
 }
 
-template<class DataType>
-int		DArray<DataType>::FreeArray( )
+template <class DataType>
+int DArray<DataType>::FreeArray()
 {
-	if( pData != NULL )
+	if (pData != NULL)
 	{
-		free( pData );
+		free(pData);
 		pData = NULL;
 	}
 
 	return 1;
 }
 
-template<class DataType>
-int		DArray<DataType>::SetValue( int k, DataType value )
+template <class DataType>
+int DArray<DataType>::SetValue(int k, DataType value)
 {
-	if( pData == NULL )
+	if (pData == NULL)
 		return 0;
 
-	if( k<0 || k>=n ) 
+	if (k < 0 || k >= n)
 		return 0;
 
-	pData[ k ] = value;
+	pData[k] = value;
 	return 1;
 }
 
-
-template<class DataType>
-int	DArray<DataType>::PrintArray()
+template <class DataType>
+int DArray<DataType>::PrintArray()
 {
-	if(n==0) 
+	if (n == 0)
 		return 0;
 
-	if( pData == NULL )
+	if (pData == NULL)
 		return 0;
 
-	for( int i=0; i<n; i++) 
+	for (int i = 0; i < n; i++)
 	{
-		char const *typen = typeid( DataType ).name();
-		if( !strcmp( typen, "double" ) )
-			printf(" %s: %lf ", typen, pData[i] );
-		else if( !strcmp( typen, "int" ) )
-			printf(" %s: %d ", typen, pData[i] );
+		char const *typen = typeid(DataType).name();
+		if (!strcmp(typen, "double"))
+			printf(" %s: %lf ", typen, (double)pData[i]);
+		else if (!strcmp(typen, "int"))
+			printf(" %s: %d ", typen, (int)pData[i]);
 
 		printf("\n");
 	}
